requirements.txt
tensorflow-gpu==2.3.0
scikit-learn==0.23.2
matplotlib==3.3.1
pandas==1.1.2
keras==2.4.0
umap-learn==0.4.6
cvxopt==1.2.5
tqdm==4.55.1
